Do not modify location in case of duplicates
When `dataset.add` is called with a dataset that already exists in the database
it used to append any new `uris` to a list of uris recorded for that dataset.
Removing this behaviour, if dataset already exists in the DB, then no changes
will be done to that record.

Higher level apis can either use `add_location` or `update` to have that
functionality. Dataset.add(ds) is a no-op if `ds` already exists in the db.